split paramgen and unittest modules
